Added a parameter to land_hash_keys to make copies of the strings.
This was preferable to a new function as quite some existing uses of the
function kept the list around after the original hash was modified (not
necessarily destroyed, removing a key was enough to invalidate a
string) - leading to hard to spot mistakes. The original way is still
available when needed, but it has to be explicitly requested now -
leading to hard to spot mistakes. The previpous way is still available
when needed, but it has to be explicitly requested now.